Final Milestone Upload

+++ Milestone Inventory/PresentationLayer/FrmInventoryList.cs
@@ -13,6 +13,7 @@ using System.Net.Quic;
 using System.Data;
 using Milestone_Inventory.BusinessLayer;
 using Milestone_Inventory.Models;
+using Milestone_Inventory.PresentationLayer;
 
 /*
  * Harlee Zebley
@@ -23,6 +24,8 @@ using Milestone_Inventory.Models;
  * This is my own work, aided by the study and use of codes learned in textbook and activities
  * 1BestCsharp blog. (2019, August 15). C# - How To Import Text File Data Into Datagridview
  * in C# [With  Source Code]. https://www.youtube.com/watch?v=Vhbv0Zvkpts
+ * No Author.(No Date). MessageBoxButtons Enum | Microsoft Docs
+ * https://learn.microsoft.com/en-us/dotnet/api/system.windows.forms.messageboxbuttons?view=windowsdesktop-9.0
  * */
 namespace Milestone_Inventory
 {
@@ -30,8 +33,12 @@ namespace Milestone_Inventory
     {
         //Declare and Initialize class variables
         List<InvItem> invItems = new List<InvItem>();
+        //list for searched items
+        List<InvItem> itemSearch = new List<InvItem>();
+        //List for edited items
+        List<InvItem> editItem = new List<InvItem>();
         private int SelectedGridIndex { get; set; }
-  
+
         public FrmInventoryList()
         {
             InitializeComponent();
@@ -144,16 +151,23 @@ namespace Milestone_Inventory
         /// <param name="e"></param>
         private void BtnDelete_ClickEventHandler(object sender, EventArgs e)
         {
-            //Instantiate the Inventory Class
-            Inventory deleteItem = new Inventory();
-            //Pass parameters to Inventory Class method
-            invItems = deleteItem.DeleteInvItem(invItems, SelectedGridIndex);
-            //Clear the DataGridView
-            gvInventoryList.DataSource = null;
-            //Read the new text into invItems
-            gvInventoryList.DataSource = invItems;
-            //Refresh the DataGridView
-            gvInventoryList.Refresh();
+            //Confirm that user would like to delete item
+            DialogResult result = MessageBox.Show("Are you sure you would like to delete this item?",
+                "Confirmation", MessageBoxButtons.YesNo);
+
+            if (result == DialogResult.Yes)
+            {
+                //Instantiate the Inventory Class
+                Inventory deleteItem = new Inventory();
+                //Pass parameters to Inventory Class method
+                invItems = deleteItem.DeleteInvItem(invItems, SelectedGridIndex);
+                //Clear the DataGridView
+                gvInventoryList.DataSource = null;
+                //Read the new text into invItems
+                gvInventoryList.DataSource = invItems;
+                //Refresh the DataGridView
+                gvInventoryList.Refresh();
+            }
         }
 
         /// <summary>
@@ -175,5 +189,68 @@ namespace Milestone_Inventory
             gvInventoryList.Refresh();
         }
 
+        /// <summary>
+        /// Search for inventory item
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void BtnSearch_ClickEventHandler(object sender, EventArgs e)
+        {
+            //declare variable to hold text
+            string searchCriteria = txtSearch.Text;
+            //instantiate Inventory class
+            Inventory searchInventory = new Inventory();
+            //pass to inventory class to execute logic
+            itemSearch = searchInventory.SearchInventory(invItems, itemSearch, searchCriteria);
+            //Send list to SearchResults form
+            FrmSearchResults frmSearchResults = new FrmSearchResults(itemSearch);
+            //Show second form
+            frmSearchResults.ShowDialog();
+            //clear the textbox
+            txtSearch.Text = "";
+        }
+
+        /// <summary>
+        /// Sort Items in DataGridView Alphabetically
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void BtnSort_ClickEventHandler(object sender, EventArgs e)
+        {
+            //Instantiate Inventory Class
+            Inventory sortedInventory = new Inventory();
+            //pass parameters to inventory class
+            invItems = sortedInventory.SortInventory(invItems);
+            //refresh DataGridView
+            gvInventoryList.DataSource = null;
+            gvInventoryList.DataSource = invItems;
+            gvInventoryList.Refresh();
+        }
+
+        /// <summary>
+        /// Edit Selected Item in DataGridView
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void BtnEdit_ClickEventHandler(object sender, EventArgs e)
+        {
+            //Instantiate the Inventory Class
+            Inventory editInvItem = new Inventory();
+            //Pass parameters Inventory Class method
+            editItem = editInvItem.EditItem(invItems, editItem, SelectedGridIndex);
+            //Open new window to display selected item information to edit
+            FrmEditItem frmEditItem = new FrmEditItem(editItem, invItems);
+            frmEditItem.ShowDialog();
+            //Instantiate Inventory Class
+            Inventory updatedInventory = new Inventory();
+            //Pass parameters to Inventory Class method
+            invItems = updatedInventory.ReadInventory(invItems);
+            //clear the DataGridView
+            gvInventoryList.DataSource = null;
+            //Get new invItems
+            gvInventoryList.DataSource = invItems;
+            //Refresh the DataGridView
+            gvInventoryList.Refresh();
+        }
     }
 }
